requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
streamlit>=1.10.0
pillow>=8.2.0
matplotlib>=3.4.2
numpy>=1.20.3
datasets>=1.8.0
transformers>=4.5.1
huggingface-hub>=0.0.12
scikit-learn>=0.24.2
seaborn>=0.11.2
watchdog>=2.1.0
opencv-python>=4.8.0
flask>=2.3.0